Compiling non-win32 was failing due to macro def
The way that the NETIMGUI_WINSOCKET_ENABLED macro was defined was failing to parse when using Clang on Windows to target an Android device. I'm not really a macro expert so I'm not sure exactly what why the other way wasn't working. Something to do with _WIN32 not being defined, as opposed to the code expecting it to be 0. Switching it to a more direct approach seems to do the trick.