Make strncpy work cross platform (#5)
The way it was set up was with defines based on compiler type. This didn't work for my use case because the #if in StringCopy() checked for (__clang__ || MSVC). In my use on Android with Clang, the former is defined. I'm guessing that whatever standard library you're using for your target console is also Clang but that version happens to have support for the _s functions, which mine doesn't.

The two options are:
1.) Find the right set of defines to make this work where each platform does its own thing
2.) Turn off the warning for MSVC and just use strncpy

Option 1 suffers from the problem that the code is different between platforms, which is non-optimal. Most sources online (e.g. https://randomascii.wordpress.com/2013/04/03/stop-using-strncpy-already/) just say to disable the warning and make your string copy function work the way you want. So this commit goes with option 2.

Co-authored-by: Aaron Lieberman <alieberman@fb.com>